secound commit

+++ respostas-exercicio1/resposta8.js
@@ -2,9 +2,9 @@ let valor1 = prompt('Digite o valor 1');
 let valor2 = prompt('Digite o valor 2');
 let valor3 = prompt('Digite o valor 3');
 
-valor1 = parseInt(valor1);
-valor2 = parseInt(valor2);
-valor3 = parseInt(valor3);
+valor1 = parseFloat(valor1);
+valor2 = parseFloat(valor2);
+valor3 = parseFloat(valor3);
 
 let multiplicacao = valor1 * valor2 * valor3;
